requirements.txt
discord.py
asyncio
dropbox
